requirements.txt
# Fake News Detection System Requirements

# Core dependencies
streamlit==1.28.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

# Text extraction
newspaper3k==0.2.8
beautifulsoup4==4.12.2
lxml==4.9.3
lxml_html_clean==0.4.3

# Web scraping and requests
requests==2.31.0

# NLP and preprocessing
nltk==3.8.1

# Pickle (included in Python standard library, but listing for clarity)
# pickle - standard library

# Optional: For Jupyter notebooks (if needed for development)
# jupyter==1.0.0
# notebook==7.0.6
